Fixes for deploying to Vercel
See explanation in https://github.com/TooTallNate/vercel-deno/issues/21.
