--- package.json
@@ -1,24 +1,0 @@
-{
-  "name": "download-your-travelmap",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "deno test src --unstable --allow-read --allow-net",
-    "fmt": "deno fmt src",
-    "build": "deno bundle ./src/index.ts ./api/index.js -c tsconfig.json",
-    "login": "npm run vercel -- login",
-    "vercel": "$(npm bin)/vercel",
-    "deploy": "npm run vercel -- deploy"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@christianhaller/npm-deno": "^1.0.0",
-    "prettier": "^2.0.5",
-    "typescript": "^3.9.6",
-    "typescript-deno-plugin": "^1.31.0",
-    "vercel": "^19.1.2"
-  }
-}
